pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.1"
description = ""
authors = ["Oscar Bahamonde <107950590+obahamonde@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.88.0"
boto3 = "^1.26.28"
python-multipart = "^0.0.5"
prisma = "^0.7.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
orjson = "^3.8.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.15.8"
black = "^22.12.0"
autopep8 = "^2.0.0"
uvicorn = {extras = ["standar"], version = "^0.20.0"}
aiohttp = "^3.8.3"
bs4 = "^0.0.1"
dnspython = "^2.2.1"
geocoder = "^1.38.1"
python-nmap = "^0.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
